src/backend/requirements.txt
# Django==3.2.9
# djangorestframework==3.12.4
# djangorestframework-simplejwt==4.4.0
# PyJWT==1.7.1
# djoser==2.0.3

Django==3.2.23
djangorestframework
#djangorestframework-simplejwt
PyJWT

psycopg2-binary

django-cors-headers
Pillow

pyotp
qrcode



dj-database-url>=0.5.0


# oauth2.0
requests
python-dotenv
